Midterm/Task2/stumgmt2/requirements.txt
fastapi
uvicorn
pydantic
pyyaml